Move Wayland client logic into WS::BaseBackend, WS::BaseTarget
Move the Wayland client logic from EGL rendering backends and targets into
standalone WS::BaseBackend and WS::BaseTarget classes, leaving only the
EGL specifics inside the Backend and Target classes implementing the
EGL-oriented renderer_backend_egl and renderer_backend_egl_target interfaces.
This makes the Wayland client code possible to reuse for backends and targets
targeting other rendering APIs.

Two notable changes in behavior:
  - The wl_display object is now destroyed through wl_display_disconnect().
  - We destroy the wl_registry object as soon as it's not necessary anymore.
