+++ CMakeLists.txt
@@ -88,6 +88,7 @@ set(WPEBACKEND_FDO_SOURCES
     src/view-backend-exportable-fdo.cpp
     src/view-backend-exportable-private.cpp
     src/ws.cpp
+    src/ws-client.cpp
 
     src/linux-dmabuf/linux-dmabuf.cpp
     src/linux-dmabuf/linux-dmabuf-protocol.c
